+++ src/livecodes/toolspane/test-viewer.ts
@@ -1,6 +1,7 @@
+/* eslint-disable import/no-internal-modules */
 import type { Config, Editors, EventsManager, TestResult, TestViewer } from '../models';
+import { sandboxService } from '../services/sandbox';
 import { getToolspaneElement } from '../UI';
-// eslint-disable-next-line import/no-internal-modules
 import * as icons from '../UI/icons';
 
 export const createTestViewer = (
@@ -62,6 +63,9 @@ export const createTestViewer = (
       if (item.classList.contains('fail')) {
         item.classList.remove('fail');
       }
+      if (item.classList.contains('skip')) {
+        item.classList.remove('skip');
+      }
     });
     const testSummary = testResultsElement.querySelector<HTMLElement>('.test-summary');
     if (testSummary) {
@@ -98,6 +102,7 @@ export const createTestViewer = (
       item.classList.add('test-result', result.status);
       result.errors
         .map((err) => err.split('at Object.<anonymous>')[0]?.trim())
+        .map((err) => err.split(`at ${sandboxService.getResultUrl()}`)[0]?.trim())
         .map((err) =>
           err.startsWith('AssertionError: ') ? err.replace('AssertionError: ', '') : err,
         )
@@ -112,6 +117,7 @@ export const createTestViewer = (
 
     const passed = results.filter((r) => r.status === 'pass').length;
     const failed = results.filter((r) => r.status === 'fail').length;
+    const skipped = results.filter((r) => r.status === 'skip').length;
     const total = results.length;
     const duration = results.reduce((totalDuration, r) => totalDuration + r.duration, 0) / 1000;
     const summary = document.createElement('div');
@@ -119,6 +125,7 @@ export const createTestViewer = (
     summary.innerHTML = `
     Tests: ${failed !== 0 ? '<span class="fail">' + failed + ' failed</span>,' : ''}
            ${passed !== 0 ? '<span class="pass">' + passed + ' passed</span>,' : ''}
+           ${skipped !== 0 ? '<span class="skip">' + skipped + ' skipped</span>,' : ''}
            ${total} total <br />
     Time: ${duration}s
 `;